[feat]복구기능 추가 message close #242 (#244)
* [feat]복구기능 추가 message close #242

* [feat]복구기능 추가 message close #242

* Update PostControllerImpl.java

게시글 삭제 및 수정 시
일반 유저이면서 작성자가 아닌 경우
throw new BusinessException("{invalid.request}"); 에러를 리턴할 수 있도록 변경했습니다!

+++ server/api/src/main/java/com/blind/api/domain/comment/controller/CommentControllerImpl.java
@@ -53,11 +53,12 @@ public class CommentControllerImpl implements CommentController {
 
         if (roleType == RoleType.USER && user.getId() != comment.getAuthorId())
             return;
-        else if (user.getId() == comment.getAuthorId())
+        if (user.getId() == comment.getAuthorId() && comment.getIsDel() == 0)
             commentService.delete(comment, RoleType.USER.getValue());
-        else
+        else if (comment.getIsDel() == 0)
             commentService.delete(comment, roleType.getValue());
-        postService.updateComment(commentService.findCommentById(commentId).getPost().getId(), -1L);
+        if (comment.getIsDel() == 0)
+            postService.updateComment(commentService.findCommentById(commentId).getPost().getId(), -1L);
     }
 
     @RequestMapping(value={"/mypage/comment"}, method = RequestMethod.GET)
@@ -65,6 +66,8 @@ public class CommentControllerImpl implements CommentController {
         Long userId = tokenService.findUserByAccessToken(HeaderUtil.getAccessToken(request)).getId();
         return commentService.findCommentByIdIn(userId, pageable);
     }
+
+
 
     RoleType setRoleType(User user, Board board) {
         if (user.getRoleType() == RoleType.ADMIN)
